mcp_client/requirements.txt
pydantic-ai==0.1.0
python-dotenv==1.1.0
asyncio==3.4.3
logfire==3.14.0